% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_rune_data.R
\name{demo_rune_data}
\alias{demo_rune_data}
\title{Generate a Small Demo Seurat Object}
\usage{
demo_rune_data(n_genes = 200, n_cells = 50)
}
\arguments{
\item{n_genes}{number of genes (rows).}

\item{n_cells}{number of cells (columns).}
}
\value{
A Seurat object containing a simulated sparse counts matrix.
}
\description{
Creates a lightweight synthetic Seurat object for testing purposes
Generated by sampling a poisson distribution
}
\examples{
demo <- demo_rune_data()
}
