% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summonData.R
\name{loadRune}
\alias{loadRune}
\title{Adds in a single dataset as a seurat object}
\usage{
loadRune(path, data = NULL)
}
\arguments{
\item{path}{filepath of data to be loaded}

\item{data}{whatever object the other already added datasets are stored in}
}
\value{
List of seurat objects with new one included
}
\description{
Adds in a single dataset as a seurat object
}
\examples{
\dontrun{
data_list <- loadRune("path")
}
}
